+++ CHANGES.md
@@ -1,3 +1,9 @@
+## 0.23.0 -- UNRELEASED
+
+Added `com.walmartlabs.lacina.resolve/wrap-resolver-result` which makes it easier
+to wrap an existing resolver function, but safely manipulate the resolved
+value.
+
 ## 0.22.1 -- 27 Oct 2015
 
 Fixes a bug that prevented operation names from working when a query